requirements.txt
cx_Oracle==8.2.1
flask-restful==0.3.9
python-dotenv==0.19.0
